perf(config): import a port without importing pydantic
Four modules in the CLI import chain reach for `suzent.config`, and all four
want only `DEFAULT_PORT` or `RUNTIME_DIR`. None of them touch `CONFIG`. Yet the
package `__init__` defined the whole configuration model, so asking for a
constant imported pydantic, the permissions schema and the file loader --
36.9ms of the 102.4ms it took to import the CLI, and the only reason pydantic
loaded at all on that path.

The constants and directory helpers move to `suzent.config.paths`, which
imports no pydantic; the model, its loader and `CONFIG` move to
`suzent.config.model`; and the package `__init__` becomes a lazy re-export
shim, the same PEP 562 pattern already used by `suzent.tools`. The ~120
existing `from suzent.config import ...` call sites are unchanged.

    import suzent.cli      102.4ms -> 65.5ms
    import suzent.config    64.2ms ->  3.4ms
    suzent --help             161ms ->  123ms

The startup import guard now covers pydantic, so this cannot quietly regress.

The legacy data-directory migration still runs when the paths module is
imported. Moving it off the import path changes when it runs, which is a
behavioral question rather than a cost one, and is left alone here.
